refactor(airbender): address PR review — trim cli_utils, simplify, add clippy
- cli_utils: trim to `init_tracing` (+ sentry); move the batch-corpus loaders
  (`BatchInputFile`/`load_batch` + gzip/hex helpers) into the integration test,
  so the production binary no longer links flate2/bincode through this crate.
  Drop the unused `resolve_batch_inputs` and the stale "messy file" TODO.
- main.rs: `--snark-vk` now defaults to the vendored `vks/snark_vk.json`
  (`default_snark_vk_path()`), mirroring `--fri-vk`; drop the Option/required_if.
- Collapse `ProofKind` into `zksync_prover_metrics::ProofType` (add Copy +
  Display there), removing the parallel enum and its `From` conversion.
- client.rs: one reqwest client with per-request timeouts (poll vs submit)
  instead of two; exponential backoff (100ms→cap 5s) on submit retries.
- worker.rs: `panic_payload_message` takes `&(dyn Any + Send)` (clippy borrowed_box).
- CI lane: add `cargo clippy --no-default-features --all-targets -- -D warnings`.

Verified CUDA-free: clippy clean, 3 unit tests pass, 4 integration tests ignored.
