+++ airbender_prover_server/crates/bin/prover_server/Cargo.toml
@@ -32,9 +32,10 @@ zksync_prover_metrics.workspace = true
 [dev-dependencies]
 # Test-only deps for the (GPU/corpus-gated, `#[ignore]`d) integration test, which
 # spawns the built binary and serves batches via a local axum mock of the data
-# handler. The verifier/host crates it also uses are already normal dependencies
-# and are available to integration tests without re-listing here.
+# handler. The verifier/host crates and bincode it also uses are already normal
+# dependencies and are available to integration tests without re-listing here.
 axum.workspace = true
+flate2.workspace = true
 hex.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "net", "time", "sync"] }
