+++ airbender_prover_server/crates/lib/cli_utils/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "zksync_cli_utils"
-description = "Shared command-line helpers for repository-owned batch input workflows"
+description = "Shared tracing initialization for the Airbender prover server"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
@@ -12,14 +12,10 @@ categories.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-bincode = { version = "2", features = ["serde"] }
-flate2.workspace = true
 sentry = { workspace = true, optional = true }
-tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter", "json"] }
-zksync_airbender_verifier.workspace = true
 
 [features]
-# Adds the Sentry tracing layer to `init_tracing`. Enabled by the server binary;
-# left off for host and the verifier libraries so they don't link Sentry.
+# Adds the Sentry tracing layer to `init_tracing`. Enabled by the server binary
+# so it reports errors/panics to Sentry; left off for any non-server consumer.
 sentry = ["dep:sentry"]
